diagnostics logic moved into separate module
